Rename manga "hooks" to "useRefreshManga"
