+++ src/screens/Manga.tsx
@@ -15,7 +15,7 @@ import { isNetworkRequestInFlight } from '@apollo/client/core/networkStatus';
 import { requestManager } from '@/lib/requests/RequestManager.ts';
 import { NavBarContext, useSetDefaultBackTo } from '@/components/context/NavbarContext';
 import { ChapterList } from '@/components/chapter/ChapterList.tsx';
-import { useRefreshManga } from '@/components/manga/hooks';
+import { useRefreshManga } from '@/components/manga/useRefreshManga.ts';
 import { MangaDetails } from '@/components/manga/MangaDetails';
 import { MangaToolbarMenu } from '@/components/manga/MangaToolbarMenu';
 import { EmptyView } from '@/components/util/EmptyView';
